Added Cards for Chronological Chao
Just added the new cards for this OP, there were two symbols that need to be added, 1. Disabled Token, and 2. Blank die result. I have other symbols in the SRC folder and will add it there but I do not know what to do to add these new symbols to the CSS and or Javascript.