Card Restriction Function
Added Icons for a number of Restritctions to equiping cards. Factions, Arc, Hull, OnePerShip, and others as I learn of them. There is likely a easyer way to do this but this was the first way I found to work.

If anyone can find a way to correct the size and shape of them, as well as adding text to the Hull icon. You would be my hero.