Many Updates
The (secondArc: "45",) is now working where it stays in place when changing scale. Added the new Ferengi/Klingon ship and edited the borg octahedron to show the correct upgrade bar and now has a unique ID

I have also been working on getting a new section online that I will be sharing more of in the issues.